Worked on example startup scripts, updated README.md

+++ scripts/startup-palmetto.sh
@@ -1,6 +1,14 @@
 #!/bin/bash
 # Startup script for Palmetto environment.
 
+# parse command-line arguments
+if [[ $# == 1 ]]; then
+    BACKEND="$1"
+else
+	echo "usage: $0 <backend>"
+	exit -1
+fi
+
 # load modules
 module purge
 module load anaconda3/5.1.0-gcc/8.3.1
@@ -10,16 +18,18 @@ module load nextflow/20.07.1
 source activate nextflow-api
 
 # start mongodb server
-killall mongod
+if [[ ${BACKEND} == "mongo" ]]; then
+    killall mongod
 
-mongod \
-    --fork \
-    --dbpath /mongo/${USER}/data \
-    --logpath /mongo/${USER}/mongod.log \
-    --bind_ip_all
+    mongod \
+        --fork \
+        --dbpath /mongo/${USER}/data \
+        --logpath /mongo/${USER}/mongod.log \
+        --bind_ip_all
+fi
 
 # start web server
 export NXF_EXECUTOR="pbspro"
 export TF_CPP_MIN_LOG_LEVEL="3"
 
-bin/server.py --backend=mongo
+bin/server.py --backend=${BACKEND}
